refactor: DIP 인터페이스 위치 통일 — OAuth 클라이언트 인터페이스를 service/auth/로 이동
AppleOAuthClient, GoogleOAuthClient 인터페이스가 service/application/ 패키지에
위치하여 DIP 인터페이스 위치가 모듈마다 불일관했던 문제를 수정.

모든 DIP 인터페이스를 service/{domain}/ 패키지에 배치하는 규칙으로 통일:
- AppleOAuthClient: service/application/ → service/auth/
- GoogleOAuthClient: service/application/ → service/auth/
- FileUploadClient: service/storage/ (기존 유지)
- TokenManager, EmailService: service/auth/ (기존 유지)
