+++ server/api/src/main/java/moment/auth/infrastructure/GoogleAuthClient.java
@@ -4,7 +4,7 @@ import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import moment.auth.dto.google.GoogleAccessToken;
 import moment.auth.dto.google.GoogleUserInfo;
-import moment.auth.service.application.GoogleOAuthClient;
+import moment.auth.service.auth.GoogleOAuthClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
