+++ server/api/src/test/java/moment/auth/service/application/AppleAuthServiceTest.java
@@ -12,6 +12,7 @@ import java.util.Optional;
 import moment.auth.domain.RefreshToken;
 import moment.auth.domain.Tokens;
 import moment.auth.dto.apple.AppleUserInfo;
+import moment.auth.service.auth.AppleOAuthClient;
 import moment.auth.service.auth.TokensIssuer;
 import moment.user.domain.ProviderType;
 import moment.user.domain.User;
